Move windows/moments rolling tests for approproate files
